net: hns3: Optimization of some variable types
driver inclusion
category: bugfix
bugzilla: NA
CVE: NA

This patch optimizs some variable types to remove the static check
warning.

Feature or Bugfix:Bugfix

+++ drivers/net/ethernet/hisilicon/hns3/hns3pf/hclge_main.h
@@ -881,7 +881,7 @@ struct hclge_vport {
 
 	u16 used_umv_num;
 
-	int vport_id;
+	u16 vport_id;
 	struct hclge_dev *back;  /* Back reference to associated dev */
 	struct hnae3_handle nic;
 	struct hnae3_handle roce;
