+++ drivers/net/ethernet/hisilicon/hns3/hns3pf/hclge_mdio.c
@@ -46,7 +46,7 @@ static int hclge_mdio_write(struct mii_bus *bus, int phyid, int regnum,
 	hnae3_set_field(mdio_cmd->phyid, HCLGE_MDIO_PHYID_M,
 			HCLGE_MDIO_PHYID_S, phyid);
 	hnae3_set_field(mdio_cmd->phyad, HCLGE_MDIO_PHYREG_M,
-			HCLGE_MDIO_PHYREG_S, regnum);
+			HCLGE_MDIO_PHYREG_S, (u32)regnum);
 
 	hnae3_set_bit(mdio_cmd->ctrl_bit, HCLGE_MDIO_CTRL_START_B, 1);
 	hnae3_set_field(mdio_cmd->ctrl_bit, HCLGE_MDIO_CTRL_ST_M,
@@ -84,7 +84,7 @@ static int hclge_mdio_read(struct mii_bus *bus, int phyid, int regnum)
 	hnae3_set_field(mdio_cmd->phyid, HCLGE_MDIO_PHYID_M,
 			HCLGE_MDIO_PHYID_S, phyid);
 	hnae3_set_field(mdio_cmd->phyad, HCLGE_MDIO_PHYREG_M,
-			HCLGE_MDIO_PHYREG_S, regnum);
+			HCLGE_MDIO_PHYREG_S, (u32)regnum);
 
 	hnae3_set_bit(mdio_cmd->ctrl_bit, HCLGE_MDIO_CTRL_START_B, 1);
 	hnae3_set_field(mdio_cmd->ctrl_bit, HCLGE_MDIO_CTRL_ST_M,
